feat(pendingRegistration): Move get all registrations function into separate file, move date formatting into utils
